+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway.hs
@@ -40,9 +40,9 @@ import Cardano.Ledger.Conway.Governance (
 import Cardano.Ledger.Conway.PParams (THKD (..))
 import Cardano.Ledger.Conway.Rules (ConwayGovPredFailure, EnactSignal, RatifyState)
 import Cardano.Ledger.Conway.Tx (AlonzoTx)
-import Cardano.Ledger.Conway.TxCert (ConwayGovCert, ConwayTxCert)
+import Cardano.Ledger.Conway.TxCert (ConwayDelegCert, ConwayGovCert, ConwayTxCert)
 import Cardano.Ledger.Credential (Credential)
-import Cardano.Ledger.Crypto (StandardCrypto)
+import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
 import Cardano.Ledger.Keys (KeyRole (..))
 import Cardano.Ledger.TxIn (TxId)
 import Constrained
@@ -74,6 +74,9 @@ import Test.Cardano.Ledger.Constrained.Conway (
   ProposalsSplit,
   certEnvSpec,
   certStateSpec,
+  dStateSpec,
+  delegCertSpec,
+  delegEnvSpec,
   genProposalsSplit,
   govCertEnvSpec,
   govCertSpec,
@@ -387,6 +390,19 @@ instance Inject (ConwayCertExecContext era) (VotingProcedures era) where
 
 instance Era era => ToExpr (ConwayCertExecContext era)
 
+disableRegCertsDelegCert ::
+  ( IsConwayUniv fn
+  , Crypto c
+  ) =>
+  Term fn (ConwayDelegCert c) ->
+  Pred fn
+disableRegCertsDelegCert delegCert =
+  (caseOn delegCert)
+    (branch $ \_ _ -> False)
+    (branch $ \_ _ -> True)
+    (branch $ \_ _ -> True)
+    (branch $ \_ _ _ -> True)
+
 instance
   IsConwayUniv fn =>
   ExecSpecRule fn "CERT" Conway
@@ -552,3 +568,17 @@ instance IsConwayUniv fn => ExecSpecRule fn "ENACT" Conway where
     first (error "ENACT failed")
       . computationResultToEither
       $ Agda.enactStep env st sig
+
+instance IsConwayUniv fn => ExecSpecRule fn "DELEG" Conway where
+  environmentSpec _ = delegEnvSpec
+
+  stateSpec _ _ = dStateSpec
+
+  signalSpec _ env st =
+    delegCertSpec env st
+      <> constrained disableRegCertsDelegCert
+
+  runAgdaRule env st sig =
+    first (\e -> OpaqueErrorString (T.unpack e) NE.:| [])
+      . computationResultToEither
+      $ Agda.delegStep env st sig